Moved ExitpollSet to a new script
